+++ app/src/androidTest/java/org/kiwix/kiwixmobile/language/LanguageRobot.kt
@@ -85,7 +85,7 @@ class LanguageRobot : BaseRobot() {
   ) {
     repeat(retryCount) { attempt ->
       try {
-        composeTestRule.waitUntil(TestUtils.TEST_PAUSE_MS_FOR_DOWNLOAD_TEST.toLong()) {
+        composeTestRule.waitUntil(TestUtils.TEST_PAUSE_MS_FOR_DOWNLOAD_TEST) {
           composeTestRule
             .onAllNodesWithContentDescription(context.getString(string.select_language_content_description))
             .fetchSemanticsNodes()
